// src/types/color.ts
export type Color =
  | 'primary'
  | 'secondary'
  | 'accent'
  | 'exhibition-a'
  | 'exhibition-b'
  | 'info'
  | 'success'
  | 'error'
  | 'warning'
  | 'white'
  | 'gray'
  | 'orange'
  | 'yellow'
  | 'green'
  | 'blue'
  | 'gray-light'
  | 'orange-light'
  | 'yellow-light'
  | 'green-light'
  | 'blue-light'